+++ bundle/src/main/java/com/adobe/acs/commons/http/headers/impl/AbstractDispatcherCacheHeaderFilter.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -23,6 +23,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.felix.scr.annotations.Activate;
 import org.apache.felix.scr.annotations.Deactivate;
 import org.apache.sling.commons.osgi.PropertiesUtil;
+import org.osgi.framework.Constants;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.cm.ConfigurationException;
 import org.osgi.service.component.ComponentContext;
@@ -30,22 +31,11 @@ import org.osgi.service.http.whiteboard.HttpWhiteboardConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
+import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Dictionary;
-import java.util.Enumeration;
-import java.util.Hashtable;
-import java.util.Iterator;
-import java.util.List;
+import java.util.*;
 
 /**
  * Abstract class to handle standard logic for registering a Dispatcher TTL header filter.
@@ -64,14 +54,14 @@ public abstract class AbstractDispatcherCacheHeaderFilter implements Filter {
 
     /**
      * Get the value to place in the Cache-Control header.
-     * 
+     *
      * @return the value of the Cache-Control header
      */
     protected abstract String getHeaderName();
 
     /**
      * Get the value to place in the Cache-Control header.
-     * 
+     *
      * @return the value of the Cache-Control header
      */
     protected abstract String getHeaderValue();
@@ -153,9 +143,10 @@ public abstract class AbstractDispatcherCacheHeaderFilter implements Filter {
         }
 
         for (String pattern : filters) {
-            Dictionary<String, String> filterProps = new Hashtable<String, String>();
+            Dictionary<String, Object> filterProps = new Hashtable<String, Object>();
 
             log.debug("Adding filter ({}) to pattern: {}", this, pattern);
+            filterProps.put(Constants.SERVICE_RANKING, PropertiesUtil.toInteger(properties.get(Constants.SERVICE_RANKING), 0));
             filterProps.put(HttpWhiteboardConstants.HTTP_WHITEBOARD_FILTER_REGEX, pattern);
             filterProps.put(HttpWhiteboardConstants.HTTP_WHITEBOARD_CONTEXT_SELECT, "(" + HttpWhiteboardConstants.HTTP_WHITEBOARD_CONTEXT_NAME + "=*)");
             ServiceRegistration filterReg = context.getBundleContext().registerService(Filter.class.getName(), this, filterProps);
@@ -166,7 +157,7 @@ public abstract class AbstractDispatcherCacheHeaderFilter implements Filter {
     @Deactivate
     protected final void deactivate(ComponentContext context) {
 
-        for(Iterator<ServiceRegistration> it = filterRegistrations.iterator(); it.hasNext(); ) {
+        for (Iterator<ServiceRegistration> it = filterRegistrations.iterator(); it.hasNext(); ) {
             ServiceRegistration registration = it.next();
             registration.unregister();
             it.remove();